almost finish

+++ internal/usecase/post.go
@@ -5,8 +5,9 @@ import (
 	"database/sql"
 	"errors"
 	"fmt"
-	customerror "github.com/firzatullahd/blog-api/internal/model/error"
 	"time"
+
+	customerror "github.com/firzatullahd/blog-api/internal/model/error"
 
 	"github.com/firzatullahd/blog-api/internal/entity"
 	"github.com/firzatullahd/blog-api/internal/model"
@@ -42,14 +43,11 @@ func (u *Usecase) CreatePost(ctx context.Context, in *model.Post) (*entity.Post,
 
 	defer func() {
 		if err != nil {
-			_ = tx.Rollback()
+			tx.Rollback()
 		}
 	}()
 
-	// if all tags exists, continue
-	// if any tag not exists, insert missing tag
 	if len(missingTags) > 0 {
-		// insert tags
 		for _, v := range missingTags {
 			tagId, err := u.repo.InsertTag(ctx, tx, &entity.Tag{
 				Label: v,
@@ -63,7 +61,7 @@ func (u *Usecase) CreatePost(ctx context.Context, in *model.Post) (*entity.Post,
 		}
 	}
 
-	postId, err := u.repo.InsertPost(ctx, tx, &entity.Post{
+	post, err := u.repo.InsertPost(ctx, tx, &entity.Post{
 		Title:   in.Title,
 		Content: in.Content,
 	})
@@ -73,21 +71,22 @@ func (u *Usecase) CreatePost(ctx context.Context, in *model.Post) (*entity.Post,
 	}
 
 	for _, v := range tagIDs {
-		err := u.repo.InsertRPostTag(ctx, tx, postId, v)
+		err := u.repo.InsertRPostTag(ctx, tx, post.ID, v)
 		if err != nil {
 			logger.Error(ctx, logCtx, err)
 			return nil, err
 		}
 	}
 
-	tx.Commit()
+	post.Tags = in.Tags
 
-	return u.GetPost(ctx, postId)
+	return post, tx.Commit()
 }
 
 func (u *Usecase) UpdatePost(ctx context.Context, in *model.Post, id uint64, email string) error {
 	logCtx := fmt.Sprintf("%T.UpdatePost", u)
 
+	// if publish post, check user's role
 	if len(in.Status) > 0 {
 		users, err := u.repo.FindUsers(ctx, &model.FilterFindUser{
 			Email: &email,
@@ -105,32 +104,6 @@ func (u *Usecase) UpdatePost(ctx context.Context, in *model.Post, id uint64, ema
 			return customerror.ErrForbidden
 		}
 	}
-
-	//rposts, err := u.repo.FindRPostTag(ctx, model.FilterFindRPost{PostID: []uint64{id}})
-	//if err != nil {
-	//	logger.Error(ctx, logCtx, err)
-	//	return err
-	//}
-
-	tags, err := u.repo.FindTag(ctx, model.FilterFindTag{
-		Label: in.Tags,
-	})
-	if err != nil {
-		logger.Error(ctx, logCtx, err)
-		return err
-	}
-
-	originalTags := make([]string, len(tags))
-	mapTags := make(map[string]uint64)
-	for _, v := range tags {
-		originalTags = append(originalTags, v.Label)
-		mapTags[v.Label] = v.ID
-	}
-
-	fmt.Printf("MAPTAG %v\n", mapTags)
-
-	missingTags := utils.FindMissing(originalTags, in.Tags)
-	unusedTags := utils.FindMissing(in.Tags, originalTags)
 
 	tx, err := u.repo.WithTransaction()
 	if err != nil {
@@ -165,9 +138,39 @@ func (u *Usecase) UpdatePost(ctx context.Context, in *model.Post, id uint64, ema
 		return err
 	}
 
+	rposts, err := u.repo.FindRPostTag(ctx, model.FilterFindRPost{PostID: id})
+	if err != nil {
+		logger.Error(ctx, logCtx, err)
+		return err
+	}
+
+	var existingTagIDs []uint64
+	for _, v := range rposts {
+		existingTagIDs = append(existingTagIDs, v.TagID)
+	}
+
+	tags, err := u.repo.FindTag(ctx, model.FilterFindTag{
+		ID: existingTagIDs,
+	})
+	if err != nil {
+		logger.Error(ctx, logCtx, err)
+		return err
+	}
+
+	existingTags := make([]string, len(tags))
+	mapTagsID := make(map[string]uint64)
+	for _, v := range tags {
+		existingTags = append(existingTags, v.Label)
+		mapTagsID[v.Label] = v.ID
+	}
+
+	fmt.Printf("MAPTAG %v\n", mapTagsID)
+
+	missingTags := utils.FindMissing(existingTags, in.Tags)
+	unusedTags := utils.FindMissing(in.Tags, existingTags)
+
 	missingTagIDs := make([]uint64, 0)
 	if len(missingTags) > 0 {
-		// insert tags
 		for _, v := range missingTags {
 			tagId, err := u.repo.InsertTag(ctx, tx, &entity.Tag{
 				Label: v,
@@ -191,12 +194,15 @@ func (u *Usecase) UpdatePost(ctx context.Context, in *model.Post, id uint64, ema
 
 	if len(unusedTags) > 0 {
 		for _, v := range unusedTags {
-			tagId, _ := mapTags[v]
-			err = u.repo.DeleteRPostTag(ctx, tx, id, tagId)
-			if err != nil {
-				logger.Error(ctx, logCtx, err)
-				return err
+			tagId, ok := mapTagsID[v]
+			if ok {
+				err = u.repo.DeleteRPostTag(ctx, tx, id, tagId)
+				if err != nil {
+					logger.Error(ctx, logCtx, err)
+					return err
+				}
 			}
+
 		}
 	}
 
@@ -224,8 +230,9 @@ func (u *Usecase) DeletePost(ctx context.Context, id uint64) error {
 
 func (u *Usecase) GetPost(ctx context.Context, id uint64) (*entity.Post, error) {
 	logCtx := fmt.Sprintf("%T.GetPost", u)
-	post, err := u.repo.FindPost(ctx, model.FilterFindPost{
-		ID: id,
+
+	posts, err := u.repo.FindPosts(ctx, model.FilterFindPost{
+		IDs: []uint64{id},
 	})
 	if err != nil {
 		logger.Error(ctx, logCtx, err)
@@ -234,6 +241,11 @@ func (u *Usecase) GetPost(ctx context.Context, id uint64) (*entity.Post, error) 
 		}
 		return nil, err
 	}
+
+	if len(posts) == 0 {
+		return nil, customerror.ErrNotFound
+	}
+	post := posts[0]
 
 	rposts, err := u.repo.FindRPostTag(ctx, model.FilterFindRPost{PostID: id})
 	if err != nil {
@@ -255,7 +267,62 @@ func (u *Usecase) GetPost(ctx context.Context, id uint64) (*entity.Post, error) 
 	for _, tag := range tags {
 		post.Tags = append(post.Tags, tag.Label)
 	}
-	post.TagIDs = tagIDs
 
-	return post, nil
+	return &post, nil
+}
+
+func (u *Usecase) SearchPost(ctx context.Context, in model.FilterSearchPost) ([]entity.Post, int, error) {
+	logCtx := fmt.Sprintf("%T.SearchPost", u)
+	// API Search Post
+	// find tags based on labels inputted, get tag id
+	// if labels inputted empty, then get all tags.
+	// find post ids from r_post_tag based on tag id
+	// find posts, map with each tags.
+	// add pagination
+
+	tags, err := u.repo.FindTag(ctx, model.FilterFindTag{Label: in.TagLabel})
+	if err != nil {
+		logger.Error(ctx, logCtx, err)
+		return nil, 0, err
+	}
+
+	mapTags := make(map[uint64]string)
+	var tagIDs []uint64
+	for _, v := range tags {
+		tagIDs = append(tagIDs, v.ID)
+		mapTags[v.ID] = v.Label
+	}
+
+	rposts, err := u.repo.FindRPostTag(ctx, model.FilterFindRPost{TagIDs: tagIDs})
+	if err != nil {
+		logger.Error(ctx, logCtx, err)
+		return nil, 0, err
+	}
+
+	mapPosts := make(map[uint64][]string)
+	var postIDs []uint64
+	for _, v := range rposts {
+		postIDs = append(postIDs, v.PostID)
+		mapPosts[v.PostID] = append(mapPosts[v.PostID], mapTags[v.TagID])
+	}
+
+	posts, err := u.repo.FindPosts(ctx, model.FilterFindPost{
+		IDs: postIDs,
+	})
+	if err != nil {
+		logger.Error(ctx, logCtx, err)
+		return nil, 0, err
+	}
+
+	countPosts, err := u.repo.CountPost(ctx)
+	if err != nil {
+		logger.Error(ctx, logCtx, err)
+		return nil, 0, err
+	}
+
+	for i := range posts {
+		posts[i].Tags = mapPosts[posts[i].ID]
+	}
+
+	return posts, countPosts, nil
 }
